+++ packages/eslint-plugin-wonder-blocks/demo/CHANGELOG.md
@@ -1,5 +1,56 @@
 # eslint-plugin-wonder-blocks-demo
 
+## 0.4.14
+
+### Patch Changes
+
+- Updated dependencies [04522d4]
+- Updated dependencies [5b4fed3]
+    - @khanacademy/wonder-blocks-tokens@18.0.0
+    - @khanacademy/wonder-blocks-button@11.7.10
+    - @khanacademy/wonder-blocks-clickable@8.2.11
+    - @khanacademy/wonder-blocks-dropdown@10.12.4
+    - @khanacademy/wonder-blocks-form@7.6.13
+    - @khanacademy/wonder-blocks-icon@6.0.2
+    - @khanacademy/wonder-blocks-icon-button@11.5.1
+    - @khanacademy/wonder-blocks-link@10.3.12
+    - @khanacademy/wonder-blocks-typography@5.0.5
+    - @khanacademy/wonder-blocks-core@12.5.0
+
+## 0.4.13
+
+### Patch Changes
+
+- Updated dependencies [63dcc94]
+- Updated dependencies [5dd4192]
+    - @khanacademy/wonder-blocks-core@12.5.0
+    - @khanacademy/wonder-blocks-icon-button@11.5.0
+    - @khanacademy/wonder-blocks-button@11.7.9
+    - @khanacademy/wonder-blocks-clickable@8.2.10
+    - @khanacademy/wonder-blocks-dropdown@10.12.3
+    - @khanacademy/wonder-blocks-form@7.6.12
+    - @khanacademy/wonder-blocks-icon@6.0.1
+    - @khanacademy/wonder-blocks-link@10.3.11
+    - @khanacademy/wonder-blocks-typography@5.0.4
+
+## 0.4.12
+
+### Patch Changes
+
+- @khanacademy/wonder-blocks-dropdown@10.12.2
+
+## 0.4.11
+
+### Patch Changes
+
+- Updated dependencies [a1f5a09]
+    - @khanacademy/wonder-blocks-icon@6.0.0
+    - @khanacademy/wonder-blocks-button@11.7.8
+    - @khanacademy/wonder-blocks-dropdown@10.12.1
+    - @khanacademy/wonder-blocks-form@7.6.11
+    - @khanacademy/wonder-blocks-icon-button@11.4.6
+    - @khanacademy/wonder-blocks-link@10.3.10
+
 ## 0.4.10
 
 ### Patch Changes