+++ packages/wonder-blocks-birthday-picker/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @khanacademy/wonder-blocks-birthday-picker
 
+## 4.1.65
+
+### Patch Changes
+
+- Updated dependencies [04522d4]
+- Updated dependencies [5b4fed3]
+    - @khanacademy/wonder-blocks-tokens@18.0.0
+    - @khanacademy/wonder-blocks-dropdown@10.12.4
+    - @khanacademy/wonder-blocks-icon@6.0.2
+    - @khanacademy/wonder-blocks-layout@3.1.62
+    - @khanacademy/wonder-blocks-typography@5.0.5
+    - @khanacademy/wonder-blocks-core@12.5.0
+
+## 4.1.64
+
+### Patch Changes
+
+- Updated dependencies [63dcc94]
+    - @khanacademy/wonder-blocks-core@12.5.0
+    - @khanacademy/wonder-blocks-dropdown@10.12.3
+    - @khanacademy/wonder-blocks-icon@6.0.1
+    - @khanacademy/wonder-blocks-layout@3.1.61
+    - @khanacademy/wonder-blocks-typography@5.0.4
+
+## 4.1.63
+
+### Patch Changes
+
+- @khanacademy/wonder-blocks-dropdown@10.12.2
+
+## 4.1.62
+
+### Patch Changes
+
+- Updated dependencies [a1f5a09]
+    - @khanacademy/wonder-blocks-icon@6.0.0
+    - @khanacademy/wonder-blocks-dropdown@10.12.1
+
 ## 4.1.61
 
 ### Patch Changes