+++ packages/wonder-blocks-clickable/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @khanacademy/wonder-blocks-clickable
 
+## 8.2.11
+
+### Patch Changes
+
+- Updated dependencies [04522d4]
+- Updated dependencies [5b4fed3]
+    - @khanacademy/wonder-blocks-tokens@18.0.0
+    - @khanacademy/wonder-blocks-core@12.5.0
+
+## 8.2.10
+
+### Patch Changes
+
+- Updated dependencies [63dcc94]
+    - @khanacademy/wonder-blocks-core@12.5.0
+
 ## 8.2.9
 
 ### Patch Changes