+++ packages/wonder-blocks-icon-button/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @khanacademy/wonder-blocks-icon-button
 
+## 11.5.1
+
+### Patch Changes
+
+- Updated dependencies [04522d4]
+- Updated dependencies [5b4fed3]
+    - @khanacademy/wonder-blocks-tokens@18.0.0
+    - @khanacademy/wonder-blocks-clickable@8.2.11
+    - @khanacademy/wonder-blocks-icon@6.0.2
+    - @khanacademy/wonder-blocks-styles@0.2.54
+    - @khanacademy/wonder-blocks-typography@5.0.5
+    - @khanacademy/wonder-blocks-core@12.5.0
+
+## 11.5.0
+
+### Minor Changes
+
+- 5dd4192: Add `onKeyDown` prop to `IconButton`, composed with (not replacing) internal Space/Enter handling. Fixes `IconButtonUnstyled` bug where a caller `onKeyDown` silently overrode internal key handling.
+
+### Patch Changes
+
+- Updated dependencies [63dcc94]
+    - @khanacademy/wonder-blocks-core@12.5.0
+    - @khanacademy/wonder-blocks-clickable@8.2.10
+    - @khanacademy/wonder-blocks-icon@6.0.1
+    - @khanacademy/wonder-blocks-typography@5.0.4
+
+## 11.4.6
+
+### Patch Changes
+
+- Updated dependencies [a1f5a09]
+    - @khanacademy/wonder-blocks-icon@6.0.0
+
 ## 11.4.5
 
 ### Patch Changes