+++ packages/wonder-blocks-pill/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @khanacademy/wonder-blocks-pill
 
+## 3.1.76
+
+### Patch Changes
+
+- Updated dependencies [04522d4]
+- Updated dependencies [5b4fed3]
+    - @khanacademy/wonder-blocks-tokens@18.0.0
+    - @khanacademy/wonder-blocks-clickable@8.2.11
+    - @khanacademy/wonder-blocks-link@10.3.12
+    - @khanacademy/wonder-blocks-styles@0.2.54
+    - @khanacademy/wonder-blocks-typography@5.0.5
+    - @khanacademy/wonder-blocks-core@12.5.0
+
+## 3.1.75
+
+### Patch Changes
+
+- Updated dependencies [63dcc94]
+    - @khanacademy/wonder-blocks-core@12.5.0
+    - @khanacademy/wonder-blocks-clickable@8.2.10
+    - @khanacademy/wonder-blocks-link@10.3.11
+    - @khanacademy/wonder-blocks-typography@5.0.4
+
+## 3.1.74
+
+### Patch Changes
+
+- @khanacademy/wonder-blocks-link@10.3.10
+
 ## 3.1.73
 
 ### Patch Changes