+++ packages/wonder-blocks-switch/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @khanacademy/wonder-blocks-switch
 
+## 3.4.12
+
+### Patch Changes
+
+- Updated dependencies [04522d4]
+- Updated dependencies [5b4fed3]
+    - @khanacademy/wonder-blocks-tokens@18.0.0
+    - @khanacademy/wonder-blocks-icon@6.0.2
+    - @khanacademy/wonder-blocks-styles@0.2.54
+    - @khanacademy/wonder-blocks-core@12.5.0
+
+## 3.4.11
+
+### Patch Changes
+
+- Updated dependencies [63dcc94]
+    - @khanacademy/wonder-blocks-core@12.5.0
+    - @khanacademy/wonder-blocks-icon@6.0.1
+
+## 3.4.10
+
+### Patch Changes
+
+- Updated dependencies [a1f5a09]
+    - @khanacademy/wonder-blocks-icon@6.0.0
+
 ## 3.4.9
 
 ### Patch Changes