+++ packages/wonder-blocks-tabs/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @khanacademy/wonder-blocks-tabs
 
+## 0.5.31
+
+### Patch Changes
+
+- Updated dependencies [04522d4]
+- Updated dependencies [5b4fed3]
+    - @khanacademy/wonder-blocks-tokens@18.0.0
+    - @khanacademy/wonder-blocks-button@11.7.10
+    - @khanacademy/wonder-blocks-dropdown@10.12.4
+    - @khanacademy/wonder-blocks-icon@6.0.2
+    - @khanacademy/wonder-blocks-link@10.3.12
+    - @khanacademy/wonder-blocks-typography@5.0.5
+    - @khanacademy/wonder-blocks-core@12.5.0
+
+## 0.5.30
+
+### Patch Changes
+
+- Updated dependencies [63dcc94]
+    - @khanacademy/wonder-blocks-core@12.5.0
+    - @khanacademy/wonder-blocks-button@11.7.9
+    - @khanacademy/wonder-blocks-dropdown@10.12.3
+    - @khanacademy/wonder-blocks-icon@6.0.1
+    - @khanacademy/wonder-blocks-link@10.3.11
+    - @khanacademy/wonder-blocks-typography@5.0.4
+
+## 0.5.29
+
+### Patch Changes
+
+- @khanacademy/wonder-blocks-dropdown@10.12.2
+
+## 0.5.28
+
+### Patch Changes
+
+- Updated dependencies [a1f5a09]
+    - @khanacademy/wonder-blocks-icon@6.0.0
+    - @khanacademy/wonder-blocks-button@11.7.8
+    - @khanacademy/wonder-blocks-dropdown@10.12.1
+    - @khanacademy/wonder-blocks-link@10.3.10
+
 ## 0.5.27
 
 ### Patch Changes